Add internal region to CFG
Currently the CFG stores its edges on the heap memory region by default.
The only options are to toggle between the heap and stack memory regions.
This commit adds the option to specify an internal region to the CFG
and modifies the createEdge API to allow CFG edges to be stored on
such region. This allows finer control over the life time of the CFG
edges.

Signed-off-by: Erick Ochoa <eochoa@ualberta.ca>